Move LoginWindow into other package
