+++ src/main/java/com/github/bachelorpraktikum/dbvisualization/view/sourcechooser/LoginWindow.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.BorderPane?>
 <BorderPane fx:id="rootPane" prefHeight="130.0" prefWidth="393.0"
   xmlns:fx="http://javafx.com/fxml/1"
-  fx:controller="com.github.bachelorpraktikum.dbvisualization.view.LoginController">
+  fx:controller="com.github.bachelorpraktikum.dbvisualization.view.sourcechooser.LoginController">
   <right>
     <AnchorPane prefHeight="124.0" prefWidth="160.0" AnchorPane.rightAnchor="0.0">
       <TextField fx:id="userField" AnchorPane.topAnchor="0.0"/>
